Refactor into uploadFile and getFile service
