+++ server/src/services/fileStorage.service.ts
@@ -22,3 +22,33 @@ const storage = new Storage({
 });
 
 export const bucket = storage.bucket(process.env.CLOUD_BUCKET as string);
+
+export const uploadFile = (file: Express.Multer.File): Promise<void> => {
+  return new Promise((resolve, reject) => {
+    if (!file || !file.buffer) {
+      reject("File or file buffer is empty");
+      return;
+    }
+
+    const blob = bucket.file(file.originalname);
+    const blobStream = blob.createWriteStream();
+
+    blobStream.on("error", reject);
+
+    blobStream.on("finish", () => {
+      resolve();
+    });
+
+    blobStream.end(file.buffer);
+  });
+};
+
+export const getFileURL = async (fileName: string): Promise<string> => {
+  const options = {
+    version: "v4" as any,
+    action: "read" as any,
+    expires: Date.now() + 15 * 60 * 1000, // URL is valid for 15 minutes
+  };
+  const [url] = await bucket.file(fileName).getSignedUrl(options);
+  return url;
+};
